Split code into seperate components to make it easier to read and to style.
